+++ crates/teloxide-core/src/lib.rs
@@ -5,7 +5,7 @@
 //! asynchronous and built using [`tokio`].
 //!
 //!```toml
-//! teloxide-core = "0.11.2"
+//! teloxide-core = "0.12.0"
 //! ```
 //! _Compiler support: requires rustc 1.82+_.
 //!